Modularize scene-specific Dirt Path fix
